# .archive/reference-snapshot/sih/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Cargo": {
        "properties": {
          "quantity_t": {
            "exclusiveMinimum": 0.0,
            "maximum": 2000000.0,
            "title": "Quantity T",
            "type": "number"
          },
          "type": {
            "enum": [
              "thermal_coal",
              "coking_coal",
              "iron_ore"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "quantity_t"
        ],
        "title": "Cargo",
        "type": "object"
      },
      "CharterWindow": {
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "CharterWindow",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpportunityRequest": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "cargo": {
            "$ref": "#/components/schemas/Cargo"
          },
          "destination_port": {
            "title": "Destination Port",
            "type": "string"
          },
          "n_simulations": {
            "default": 300,
            "maximum": 2000.0,
            "minimum": 100.0,
            "title": "N Simulations",
            "type": "integer"
          },
          "origin_port": {
            "title": "Origin Port",
            "type": "string"
          },
          "required_by": {
            "format": "date",
            "title": "Required By",
            "type": "string"
          },
          "risk_aversion": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Aversion"
          },
          "vessel_class": {
            "anyOf": [
              {
                "enum": [
                  "Handysize",
                  "Supramax",
                  "Panamax",
                  "Capesize"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vessel Class"
          }
        },
        "required": [
          "cargo",
          "origin_port",
          "destination_port",
          "required_by"
        ],
        "title": "OpportunityRequest",
        "type": "object"
      },
      "OptimizeCharterRequest": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "cargo": {
            "$ref": "#/components/schemas/Cargo"
          },
          "contract_horizon_voyages": {
            "anyOf": [
              {
                "maximum": 12.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Horizon Voyages"
          },
          "deadline_miss_penalty_usd": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Miss Penalty Usd"
          },
          "destination_port": {
            "title": "Destination Port",
            "type": "string"
          },
          "include_milp_crosscheck": {
            "default": false,
            "title": "Include Milp Crosscheck",
            "type": "boolean"
          },
          "include_opportunity_score": {
            "default": true,
            "title": "Include Opportunity Score",
            "type": "boolean"
          },
          "n_simulations": {
            "anyOf": [
              {
                "maximum": 5000.0,
                "minimum": 100.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Simulations"
          },
          "origin_port": {
            "title": "Origin Port",
            "type": "string"
          },
          "required_by": {
            "format": "date",
            "title": "Required By",
            "type": "string"
          },
          "risk_aversion": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Aversion"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "required": [
          "cargo",
          "origin_port",
          "destination_port",
          "required_by"
        ],
        "title": "OptimizeCharterRequest",
        "type": "object"
      },
      "OptimizeResponse": {
        "properties": {
          "alternatives": {
            "items": {
              "type": "object"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "cost_breakdown_usd": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Cost Breakdown Usd",
            "type": "object"
          },
          "explanation": {
            "items": {
              "type": "string"
            },
            "title": "Explanation",
            "type": "array"
          },
          "idle": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Idle",
            "type": "object"
          },
          "infeasible": {
            "items": {
              "type": "object"
            },
            "title": "Infeasible",
            "type": "array"
          },
          "milp_crosscheck": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Milp Crosscheck"
          },
          "opportunity": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity"
          },
          "provenance": {
            "additionalProperties": {
              "enum": [
                "measured",
                "derived",
                "estimated",
                "simulated_demo",
                "expert_set"
              ],
              "type": "string"
            },
            "title": "Provenance",
            "type": "object"
          },
          "recommendation": {
            "$ref": "#/components/schemas/Recommendation"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "risk": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk"
          },
          "runtime_ms": {
            "title": "Runtime Ms",
            "type": "integer"
          }
        },
        "required": [
          "recommendation",
          "cost_breakdown_usd",
          "idle",
          "alternatives",
          "infeasible",
          "explanation",
          "assumptions",
          "provenance",
          "as_of",
          "runtime_ms",
          "request_id"
        ],
        "title": "OptimizeResponse",
        "type": "object"
      },
...
      "Recommendation": {
        "properties": {
          "charter_opportunity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Charter Opportunity Score"
          },
          "charter_window": {
            "$ref": "#/components/schemas/CharterWindow"
          },
          "contract_structure": {
            "title": "Contract Structure",
            "type": "string"
          },
          "cvar_90_usd": {
            "title": "Cvar 90 Usd",
            "type": "number"
          },
          "execution": {
            "title": "Execution",
            "type": "string"
          },
          "expected_cost_inr": {
            "title": "Expected Cost Inr",
            "type": "number"
          },
          "expected_cost_usd": {
            "title": "Expected Cost Usd",
            "type": "number"
          },
          "expected_usd_per_tonne": {
            "title": "Expected Usd Per Tonne",
            "type": "number"
          },
          "fx_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fx Date"
          },
          "fx_rate": {
            "title": "Fx Rate",
            "type": "number"
          },
          "interval_80_usd": {
            "items": {
              "type": "number"
            },
            "title": "Interval 80 Usd",
            "type": "array"
          },
          "p_deadline_miss": {
            "title": "P Deadline Miss",
            "type": "number"
          },
          "risk": {
            "title": "Risk",
            "type": "string"
          },
          "risk_adjusted_cost_usd": {
            "title": "Risk Adjusted Cost Usd",
            "type": "number"
          },
          "vessel_class": {
            "title": "Vessel Class",
            "type": "string"
          },
          "voyages": {
            "title": "Voyages",
            "type": "integer"
          }
        },
        "required": [
          "vessel_class",
          "voyages",
          "execution",
          "contract_structure",
          "charter_window",
          "expected_cost_usd",
          "expected_cost_inr",
          "expected_usd_per_tonne",
          "fx_rate",
          "fx_date",
          "interval_80_usd",
          "cvar_90_usd",
          "p_deadline_miss",
          "risk_adjusted_cost_usd",
          "risk"
        ],
        "title": "Recommendation",
        "type": "object"
      },
...
      "SimulateResponse": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "distributions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Distributions",
            "type": "object"
          },
          "infeasible": {
            "items": {
              "type": "object"
            },
            "title": "Infeasible",
            "type": "array"
          },
          "n_simulations": {
            "title": "N Simulations",
            "type": "integer"
          },
          "provenance": {
            "additionalProperties": {
              "enum": [
                "measured",
                "derived",
                "estimated",
                "simulated_demo",
                "expert_set"
              ],
              "type": "string"
            },
            "title": "Provenance",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/StrategyResult"
            },
            "title": "Results",
            "type": "array"
          },
          "runtime_ms": {
            "title": "Runtime Ms",
            "type": "integer"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "winner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner"
          }
        },
        "required": [
          "results",
          "infeasible",
          "winner",
          "n_simulations",
          "seed",
          "runtime_ms",
          "distributions",
          "as_of",
          "assumptions",
          "provenance",
          "request_id"
        ],
        "title": "SimulateResponse",
        "type": "object"
      },
      "SimulateStrategyRequest": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "cargo": {
            "$ref": "#/components/schemas/Cargo"
          },
          "deadline_miss_penalty_usd": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Miss Penalty Usd"
          },
          "destination_port": {
            "title": "Destination Port",
            "type": "string"
          },
          "n_simulations": {
            "default": 1000,
            "maximum": 5000.0,
            "minimum": 100.0,
            "title": "N Simulations",
            "type": "integer"
          },
          "origin_port": {
            "title": "Origin Port",
            "type": "string"
          },
          "required_by": {
            "format": "date",
            "title": "Required By",
            "type": "string"
          },
          "risk_aversion": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Aversion"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "strategies": {
            "items": {
              "type": "string"
            },
            "title": "Strategies",
            "type": "array"
          }
        },
        "required": [
          "cargo",
          "origin_port",
          "destination_port",
          "required_by"
        ],
        "title": "SimulateStrategyRequest",
        "type": "object"
      },
...
      "StrategyResult": {
        "properties": {
          "breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Breakdown",
            "type": "object"
          },
          "cost": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Cost",
            "type": "object"
          },
          "execution": {
            "title": "Execution",
            "type": "string"
          },
          "feasible": {
            "default": true,
            "title": "Feasible",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "idle_cost_usd": {
            "title": "Idle Cost Usd",
            "type": "number"
          },
          "idle_days": {
            "title": "Idle Days",
            "type": "number"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "notes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "p_deadline_miss": {
            "title": "P Deadline Miss",
            "type": "number"
          },
          "risk_adjusted_cost": {
            "title": "Risk Adjusted Cost",
            "type": "number"
          },
          "structure": {
            "title": "Structure",
            "type": "string"
          },
          "total_days": {
            "title": "Total Days",
            "type": "number"
          },
          "usd_per_tonne": {
            "title": "Usd Per Tonne",
            "type": "number"
          },
          "vessel_class": {
            "title": "Vessel Class",
            "type": "string"
          },
          "voyages": {
            "title": "Voyages",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "label",
          "vessel_class",
          "voyages",
          "structure",
          "execution",
          "total_days",
          "cost",
          "usd_per_tonne",
          "p_deadline_miss",
          "risk_adjusted_cost",
          "breakdown",
          "idle_days",
          "idle_cost_usd"
        ],
        "title": "StrategyResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nDecision-support API for chartering bulk carriers into India's East Coast ports.\n\n**Forecast the market \u2192 understand the constraints \u2192 simulate charter strategies \u2192\nrecommend the best decision.**\n\nThis is not a freight price oracle. It is a constrained decision engine that treats\nthe freight forecast as an uncertain input, not as the answer. Every response carries\n`provenance`, `as_of` and `assumptions`, and every infeasible option comes back with\nthe constraint and the source that ruled it out.\n\nProvenance values: `measured` (from an official published record), `derived` (computed\nby one of our models from measured inputs), `estimated` (a documented judgement where no\npublication exists), `expert_set` (a hand-curated parameter), `simulated_demo` (a\nseeded demo series standing in for a feed that needs credentials).\n",
    "title": "SIH26006 \u2014 Charter Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/opportunity-score": {
      "post": {
        "operationId": "opportunity_score_api_v1_opportunity_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpportunityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Charter Opportunity Score for a requirement",
        "tags": [
          "decision"
        ]
      }
    },
    "/api/v1/optimize-charter": {
      "post": {
        "operationId": "optimize_charter_api_v1_optimize_charter_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeCharterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recommend a charter strategy and explain it",
        "tags": [
          "decision"
        ]
      }
    },
...
    "/api/v1/simulate-strategy": {
      "post": {
        "operationId": "simulate_strategy_api_v1_simulate_strategy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateStrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Charter Strategy Digital Twin \u2014 compare strategies under uncertainty",
        "tags": [
          "decision"
        ]
      }
    },
...
  },
...
}